+++ src/graphics/screens/crt_test_screen.h
@@ -35,7 +35,7 @@ class CrtTestScreen : public IScreen
         std::vector<std::string> *screen_names);
     ~CrtTestScreen();
 
-    virtual void update(const float delta_time);
+    virtual void update(const double delta_time);
     virtual void draw(Renderer &renderer, bool lightweight = false);
 };
 }
